drivers: npmx: shell: move pof code to the separate file
