+++ drivers/npmx/CMakeLists.txt
@@ -54,6 +54,7 @@ if(CONFIG_NPMX_SHELL)
     zephyr_library_sources(shell/gpio.c)
     zephyr_library_sources(shell/ldsw.c)
     zephyr_library_sources(shell/led.c)
+    zephyr_library_sources(shell/pof.c)
 endif()
 
 # Get npmx include file
